change MODULE_IMPORTER determination; fix aws_codeseeder.services imports
